Create ExtensionFunctions.kt
