--- src/main/kotlin/frc/robot/lib/extensions/ChassisSpeedsExtensions.kt
@@ -1,8 +1,0 @@
-package frc.robot.lib.extensions
-
-import edu.wpi.first.math.kinematics.ChassisSpeeds
-import kotlin.math.hypot
-
-object ChassisSpeedsExtensions {
-fun ChassisSpeeds.getSpeed() = hypot(vxMetersPerSecond, vyMetersPerSecond)
-}
