--- src/main/kotlin/frc/robot/lib/extensions/ListAnyExtensions.kt
@@ -1,15 +1,0 @@
-package frc.robot.lib.extensions
-
-object ListAnyExtensions {
-    fun List<Any>.toDoubleArray(): DoubleArray {
-        return this.map { it as Double }.toDoubleArray()
-    }
-
-    fun List<Any>.toIntArray(): IntArray {
-        return this.map { it as Int }.toIntArray()
-    }
-
-    fun List<Any>.toBooleanArray(): BooleanArray {
-        return this.map { it as Boolean }.toBooleanArray()
-    }
-}
