--- src/main/kotlin/frc/robot/lib/extensions/TranslationExtensions.kt
@@ -1,9 +1,0 @@
-package frc.robot.lib.extensions
-
-import edu.wpi.first.math.geometry.Rotation2d
-import edu.wpi.first.math.geometry.Translation2d
-
-object TranslationExtensions {
-
-    fun Translation2d.getRotationToTranslation(other: Translation2d): Rotation2d = (this - other).angle
-}
